refactor(server): refactored to loose typescript
